+++ CHANGELOG.md
@@ -4,6 +4,24 @@ Please visit https://github.com/shd101wyy/vscode-markdown-preview-enhanced/relea
 
 ## [Unreleased]
 
+## [0.9.12] - 2025-03-16
+
+### Changes
+
+- Use [sharp](https://www.npmjs.com/package/sharp) to convert svg element to png file if `imageMagickPath` is empty. [crossnote#366](https://github.com/shd101wyy/crossnote/issues/366)
+
+### Updates
+
+- Updated `mermaid` version to the latest `11.5.0`.
+- Updated `katex` version to the latest `0.16.21`.
+- Updated `prismjs` version to the latest `1.30.0`.
+- Updated `bit-field` version to the latest `1.9.0`.
+
+### Bug fixes
+
+- Fixed the import the crossnote as nodejs esm module. [crossnote#357](https://github.com/shd101wyy/crossnote/issues/357)
+- Fixed a bug of using `enableExtendedTableSyntax`. [crossnote#369](https://github.com/shd101wyy/crossnote/issues/369)
+
 ## [0.9.11] - 2024-11-08
 
 ### Updates